Program: removed PumpPistons(..) use, added WIP pumping in drill/contract states.
The WIP sections are commented out, since not sure they're
necessary.

Also, it may be that pumping pistons will require updating
gear positions, since without doing so, the gears may end up
in exactly the same spot they were before pumping; and if
a check is made in rapid succession on the same gears in
different states, it may _seem_ that the gear hasn't moved
in _this_ state's tick, but the gears' locations are there
from the _previous_ state.

If so, should probably clean up the locations in "previous
use" states when they're no longer needed. But we've
gotten away with it so far...